+++ transparai.php
@@ -43,6 +43,7 @@ require_once TRANSPARAI_PLUGIN_DIR . 'includes/class-delivery.php';
 require_once TRANSPARAI_PLUGIN_DIR . 'includes/class-frontend.php';
 require_once TRANSPARAI_PLUGIN_DIR . 'includes/class-notice.php';
 require_once TRANSPARAI_PLUGIN_DIR . 'includes/class-woocommerce.php';
+require_once TRANSPARAI_PLUGIN_DIR . 'includes/class-rest.php';
 require_once TRANSPARAI_PLUGIN_DIR . 'includes/class-integrations.php';
 
 if ( is_admin() ) {
@@ -76,6 +77,7 @@ if ( ! class_exists( 'TransparAI' ) ) {
 			TransparAI_Frontend::init();
 			TransparAI_Notice::init();
 			TransparAI_WooCommerce::init();
+			TransparAI_REST::init();
 			TransparAI_Integrations::init();
 
 			if ( is_admin() ) {